+++ CHANGELOG.md
@@ -15,7 +15,15 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :rocket: New Feature
 
-* Initial release `bDummyModuleLoader`
+* Initial release:
+  * `bDummyModuleLoader`
+  * `bDummyState`
+
+#### :house: Internal
+
+* Added tests:
+  * `super/i-block/modules/module-loader`
+  * `super/i-block/modules/state`
 
 ## v3.0.0-rc.216 (2021-07-26)